# api.yml
...
swagger: '2.0'
info:
  version: 1.0.0
  title: api-activos-fijo Microservice
  description: Java Microservice + MySQL Database
  contact:
    name: Jhon alexander chacon torres
    email: jhachacont@correo.udistrital.edu.co
  license:
    name: PruebaConocimiento
host: localhost
basePath: /api-activos-fijo-ms/resources
schemes:
- http
consumes:
- application/json
produces:
- application/json
paths:
...
  /color:
    get:
      summary: retorna todos los registros de colores registrados
      operationId: getAllColor
      parameters:
      - name: firstResult
        in: query
        description: el primer resultado a retornar
        required: false
        type: integer
      - name: maxResult
        in: query
        description: el maximo numero de resultados a retornar
        required: false
        type: integer
      responses:
        200:
          description: Ok
          schema:
            type: array
            items:
              $ref: '#/definitions/Color'
        406:
          description: Not Acceptable (Invalid Params)
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'
    post:
      summary: crear nuevos colores
      operationId: createColor
      parameters:
      - name: Color
        in: body
        required: true
        description: Objeto Color a crear
        schema:
          $ref: '#/definitions/Color'
      responses:
        201:
          description: Respuesta para la creacion del objeto
          schema:
            $ref: '#/definitions/Color'
        400:
          description: Bad Request falta un dato
          schema:
            $ref: '#/definitions/ErrorModel'
        403:
          description: Forbidden el dato ya existe
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'
    put:
      summary: Actualizar un objeto de color
      operationId: updateColor
      parameters:
      - name: color
        in: body
        required: true
        description: Objeto color a actualizar
        schema:
          $ref: '#/definitions/Color'
      responses:
        200:
          description: Ok
          schema:
            $ref: '#/definitions/Color'
        400:
          description: Bad Request falta un dato
          schema:
            $ref: '#/definitions/ErrorModel'
        404:
          description: Not Found no existe el dato
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: unexpected error
          schema:
            $ref: '#/definitions/ErrorModel'
    delete:
      summary: Eliminar color segun llave o codigo
      operationId: deleteColor
      parameters:
      - name: codigo
        in: query
        required: true
        description: codigo o lleve del color
        type: string
      responses:
        200:
          description: Ok
          schema:
            $ref: '#/definitions/Color'
        400:
          description: Bad Request falta un dato
          schema:
            $ref: '#/definitions/ErrorModel'
        404:
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: unexpected error
          schema:
            $ref: '#/definitions/ErrorModel'
  /color/byCodigo:
    get:
      summary: Retorna un color segun su codigo
      operationId: getColorByCode
      parameters:
      - name: codigo
        in: query
        description: codigo del color
        required: true
        type: string
      responses:
        200:
          description: Ok
          schema:
            $ref: '#/definitions/Color'
        400:
          description: Bad Request falta un dato
          schema:
            $ref: '#/definitions/ErrorModel'
        404:
          description: Not Found no existe el dato
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: unexpected error
          schema:
            $ref: '#/definitions/ErrorModel'
  /estadoActual:
    get:
      summary: retorna todos los registros de estado actual registrados
      operationId: getAllEstadoActual
      parameters:
      - name: firstResult
        in: query
        description: el primer resultado a retornar
        required: false
        type: integer
      - name: maxResult
        in: query
        description: el maximo numero de resultados a retornar
        required: false
        type: integer
      responses:
        200:
          description: Ok
          schema:
            type: array
            items:
              $ref: '#/definitions/EstadoActual'
        406:
          description: Not Acceptable (Invalid Params)
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'
    post:
      summary: crear nuevos estados actuales
      operationId: createEstadoActual
      parameters:
      - name: estadoActual
        in: body
        required: true
        description: Objeto estado actual a crear
        schema:
          $ref: '#/definitions/EstadoActual'
      responses:
        201:
          description: Respuesta para la creacion del objeto
          schema:
            $ref: '#/definitions/EstadoActual'
        400:
          description: Bad Request falta un dato
          schema:
            $ref: '#/definitions/ErrorModel'
        403:
          description: Forbidden el dato ya existe
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorModel'
    put:
      summary: Actualizar un objeto de estado actual
      operationId: updateEstadoActual
      parameters:
      - name: estadoActual
        in: body
        required: true
        description: Objeto estado actual a actualizar
        schema:
          $ref: '#/definitions/EstadoActual'
      responses:
        200:
          description: Ok
          schema:
            $ref: '#/definitions/EstadoActual'
        400:
          description: Bad Request falta un dato
          schema:
            $ref: '#/definitions/ErrorModel'
        404:
          description: Not Found no existe el dato
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: unexpected error
          schema:
            $ref: '#/definitions/ErrorModel'
    delete:
      summary: Eliminar estado actual segun llave o codigo
      operationId: deleteEstadoActual
      parameters:
      - name: codigo
        in: query
        required: true
        description: codigo o lleve del estado actual
        type: string
      responses:
        200:
          description: Ok
          schema:
            $ref: '#/definitions/EstadoActual'
        400:
          description: Bad Request falta un dato
          schema:
            $ref: '#/definitions/ErrorModel'
        404:
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: unexpected error
          schema:
            $ref: '#/definitions/ErrorModel'
  /estadoActual/byCodigo:
    get:
      summary: Retorna un estado actual segun su codigo
      operationId: getEstadoActualByCode
      parameters:
      - name: codigo
        in: query
        description: codigo del estado actual
        required: true
        type: string
      responses:
        200:
          description: Ok
          schema:
            $ref: '#/definitions/EstadoActual'
        400:
          description: Bad Request falta un dato
          schema:
            $ref: '#/definitions/ErrorModel'
        404:
          description: Not Found no existe el dato
          schema:
            $ref: '#/definitions/ErrorModel'
        500:
          description: unexpected error
          schema:
            $ref: '#/definitions/ErrorModel'
...
definitions:
...
  Color:
    type: object
    required: ["codigo", "descripcion"]
    properties:
      codigo:
        type: string
      descripcion:
        type: string
  EstadoActual:
    type: object
    required: ["codigo", "descripcion"]
    properties:
      codigo:
        type: string
      descripcion:
        type: string
...
  ErrorModel:
    type: object
    required:
    - code
    - description
    - id
    properties:
      code:
        type: integer
      description:
        type: object
        properties:
          key:
            type: string
          value:
            type: string
      id:
        type: string